requirements.txt
Django<1.8,>=1.4.2
DjangoRestless>=0.0.10
python-subunit>=0.0.18
testtools>=0.9.30
testrepository>=0.0.18
subunit2sql>=0.7.0
